Add auto-formatting with Conform
This also includes a `ToggleFormattingOnSave` command that can be used
to disable the formatting on save.

There is a slight issue, that I can't quite figure out -> and that I
can't always toggle it *back on*, correctly.

I run the command, and it stays disabled, requiring multiple attempts to
re-enable.